change folder structure, make variable for navbarmenu, import image
