+++ pages/index.tsx
@@ -1,7 +1,7 @@
 import Head from 'next/head'
 import Image from 'next/image'
 import styles from '@/styles/Home.module.css'
-import Layout from './components/Navbar/Layout'
+import Layout from '../components/pages/Navbar/Layout'
 
 export default function Index() {
   return (
